# backend/docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "ユーザー・認証・フレンド・ブロック・アバターを扱う公開 API の仕様。\n外部クライアントから叩けるよう、認証は Cookie ではなく Bearer トークンで行う。\n現時点では仕様のみで、エンドポイントの実装はこれから追加される。",
        "title": "ft_transcendence API",
        "contact": {
            "name": "42 Transcendence"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/api/v1",
    "paths": {
...
        "/friends": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "status=accepted の関係を返す。status を指定すると申請中なども取得できる。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "友達一覧",
                "parameters": [
                    {
                        "enum": [
                            "pending",
                            "accepted",
                            "rejected"
                        ],
                        "type": "string",
                        "default": "accepted",
                        "description": "絞り込む状態",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "取得件数 (1-100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "取得開始位置",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apispec.FriendshipListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/friends/requests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "direction=incoming で自分宛て、outgoing で自分が出した申請を返す。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "友達申請の一覧",
                "parameters": [
                    {
                        "enum": [
                            "incoming",
                            "outgoing"
                        ],
                        "type": "string",
                        "default": "incoming",
                        "description": "申請の向き",
                        "name": "direction",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apispec.FriendshipListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "pending の関係を作る。相手が既に自分へ申請していれば、その場で accepted になる。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "友達申請を送る",
                "parameters": [
                    {
                        "description": "申請先",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apispec.FriendRequestCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/apispec.FriendshipResponse"
                        }
                    },
                    "400": {
                        "description": "自分自身は指定できない",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "どちらかがブロックしている",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "相手が存在しない",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "既に申請済み、または友達",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/friends/{userId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "関係を削除する。accepted なら友達解除、pending なら申請の取り消しになる。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "friends"
                ],
                "summary": "友達を解除／申請を取り消す",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "相手のユーザー ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "削除完了"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "関係が無い",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/leaderboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "レーティングの高い順に返す。自分の順位は items の外でも me に入る。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leaderboard"
                ],
                "summary": "ランキング",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "取得件数 (1-100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "取得開始位置",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apispec.LeaderboardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/matches": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "自分が参加した決着済みの対戦を新しい順に返す。日付範囲とモード・勝敗で絞り込める。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "対戦履歴",
                "parameters": [
                    {
                        "type": "string",
                        "format": "date-time",
                        "description": "この日時以降に終了した対戦",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "description": "この日時以前に終了した対戦",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ranked",
                            "casual",
                            "ai",
                            "friend"
                        ],
                        "type": "string",
                        "description": "対戦モード",
                        "name": "mode",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "win",
                            "loss",
                            "draw"
                        ],
                        "type": "string",
                        "description": "自分から見た勝敗",
                        "name": "outcome",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "取得件数 (1-100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "取得開始位置",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apispec.MatchListResponse"
                        }
                    },
                    "400": {
                        "description": "from \u003e to、または範囲外の値",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ゲーム側が決着時に呼ぶ入口。レーティング・XP・レベルの更新と実績の判定を行い、\n参加者に対して SSE で更新を通知する。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "対戦結果を記録する",
                "parameters": [
                    {
                        "description": "決着した対戦",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apispec.MatchCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "呼び出したユーザーから見た結果",
                        "schema": {
                            "$ref": "#/definitions/apispec.MatchResponse"
                        }
                    },
                    "400": {
                        "description": "参加者が 2 人でない、勝敗の組み合わせが矛盾している、終了が開始より前",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "参加者が存在しない",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/friends": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "フレンド一覧（Public API）",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apispec.FriendshipListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/friends/requests": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "フレンド申請（Public API）",
                "parameters": [
                    {
                        "description": "申請先",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apispec.FriendRequestCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/apispec.FriendshipResponse"
                        }
                    },
                    "400": {
                        "description": "friend_self",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "insufficient_scope / friend_blocked",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "user_not_found",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "friend_already_requested / already_friends",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/friends/{userId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "public"
                ],
                "summary": "フレンド解除（Public API）",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "相手のユーザー ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "friend_not_found",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/leaderboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "ランキング（Public API）",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "件数",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "開始位置",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apispec.LeaderboardResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/apispec.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apispec.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "unauthorized"
                },
                "message": {
                    "type": "string",
                    "example": "認証が必要です"
                }
            }
        },
        "apispec.FriendRequestCreateRequest": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "user_id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "3f0c1a2e-9a5b-4c7d-8e1f-2b3c4d5e6f70"
                }
            }
        },
...
        "apispec.FriendshipListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apispec.FriendshipResponse"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 20
                },
                "offset": {
                    "type": "integer",
                    "example": 0
                },
                "total": {
                    "description": "条件に合致する総件数",
                    "type": "integer",
                    "example": 128
                }
            }
        },
...
        "apispec.MatchListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apispec.MatchResponse"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 20
                },
                "offset": {
                    "type": "integer",
                    "example": 0
                },
                "total": {
                    "description": "条件に合致する総件数",
                    "type": "integer",
                    "example": 128
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "`Bearer {access_token}` 形式で指定する。トークンは POST /auth/login などで発行する。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}